requirements.txt
asgiref==3.8.1
Django==4.2.14
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==26.0.0
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pydot==3.0.1
pydotplus==2.0.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
